Eco-Scan/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pillow==10.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
bcrypt==4.0.1
pyjwt==2.8.0